Fixed issues with adding offsets to PC
Negative offsets were interpreted as unsigned, wrecking havoc on the
desired program flow.

+++ gbEmulator/coreEmulator/romState.m
@@ -68,7 +68,7 @@
 {
     PC++;
 }
-- (void) addToPC:(int)offset
+- (void) addToPC:(int8_t)offset
 {
     PC += offset;
 }
